serverless/websocket-test/client/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
python-dotenv==1.0.1
    # via -r requirements.in
websocket-client==1.8.0
    # via -r requirements.in
websockets==14.1
    # via -r requirements.in
